move helper functions out of layout
