+++ src/helper/getModule.js
@@ -1,9 +1,3 @@
-import React, { Fragment } from 'react';
-import PropTypes from 'prop-types';
-import Path from 'path';
-import Fs from 'fs';
-
-
 /**
  * Gets the relevant module
  *
