+++ src/layout/component/header.js
@@ -1,7 +1,8 @@
 import ComponentStatus from './status';
-import GetModule from './../getModule';
+import GetModule from './../../helper/getModule';
+import GetData from './../../helper/getData';
+
 import PropTypes from 'prop-types';
-import GetData from './../getData';
 import React from 'react';
