+++ src/layout/component/status.js
@@ -1,7 +1,7 @@
 import AUlinkList                               from '../../_uikit/layout/link-list';
 import AUheading                                from '../../_uikit/layout/headings';
 import Contributors                             from './contributors';
-import GetData, { GetComponentValue, GetState } from '../getData';
+import GetData, { GetComponentValue, GetState } from '../../helper/getData';
 
 import React, { Fragment }                      from 'react';
 import PropTypes                                from 'prop-types';
